CTCP-5580: Move goods reference into new equipment helper
