+++ test/base/SpecBase.scala
@@ -44,6 +44,7 @@ trait SpecBase extends AnyFreeSpec with Matchers with OptionValues with TryValue
   val departureIdP5: String    = "643cffea2dca70b2"
   val arrivalIdP5: String      = "62f4ebbb765ba8c2"
 
+  val index: Index                   = Index(0)
   val incidentIndex: Index           = Index(0)
   val transportEquipmentIndex: Index = Index(0)
