+++ test/helper/IncidentP5TransportEquipmentHelperSpec.scala
@@ -17,7 +17,9 @@
 package helper
 
 import base.SpecBase
+import generated.GoodsReferenceType01
 import generators.Generators
+import org.scalacheck.Arbitrary.arbitrary
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import utils.IncidentP5TransportEquipmentHelper
 import viewModels.sections.Section.AccordionSection
@@ -40,6 +42,26 @@ class IncidentP5TransportEquipmentHelperSpec extends SpecBase with ScalaCheckPro
         }
       }
 
+      "goodsReferenceNumber" - {
+        "must return a row" in {
+
+          forAll(arbitrary[BigInt]) {
+            referenceNumber =>
+              val transportEquipment =
+                arbitraryTransportEquipmentType07.arbitrary.sample.value.copy(
+                  GoodsReference = Seq(arbitraryGoodsReferenceType01.arbitrary.sample.value.copy(declarationGoodsItemNumber = referenceNumber))
+                )
+
+              val helper = new IncidentP5TransportEquipmentHelper(transportEquipment)
+              val result = helper.goodsReferenceNumber(referenceNumber, index).value
+
+              result.key.value mustBe s"Goods item number ${index.display}"
+              result.value.value mustBe referenceNumber.toString()
+              result.actions must not be defined
+          }
+        }
+      }
+
     }
 
     "sections" - {
@@ -55,6 +77,20 @@ class IncidentP5TransportEquipmentHelperSpec extends SpecBase with ScalaCheckPro
           result.rows.size mustBe 1
           result.children.size mustBe 0
           result.isOpen mustBe true
+        }
+      }
+
+      "goodsReferenceSection" - {
+        "must return a accordion section with goodsReference rows" in {
+          val transportEquipment = arbitraryTransportEquipmentType07.arbitrary.sample.value.copy(
+            GoodsReference = Seq(GoodsReferenceType01("1", 1), GoodsReferenceType01("2", 2))
+          )
+
+          val helper = new IncidentP5TransportEquipmentHelper(transportEquipment)
+          val result = helper.goodsReferenceSection
+
+          result mustBe a[AccordionSection]
+          result.rows.size mustBe 2
         }
       }
     }
